oldarasaac/upgrade/requirements.txt
backcall==0.1.0
decorator==4.4.1
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
mongoengine==0.18.2
mysqlclient==1.4.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptvsd==4.3.2
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.9.0
python-dotenv==0.10.3
six==1.12.0
traitlets==4.3.3
wcwidth==0.1.7